improved error handling when there are issues fetching videos

+++ src/store-video-details/store-all-video-details.js
@@ -9,33 +9,46 @@ export const storeAllVideoDetails = async (channels) => {
   for (const channel of channels) {
     logger.info(channel.channel_title);
 
-    const videos = videoApiResponseToDbColumns(
-      await getAllPlaylistAsmrVideos(channel.upload_playlist_id)
-    );
+    const videos = await getAllPlaylistAsmrVideos(channel.upload_playlist_id);
+    if (!videos) continue; // skip if there was an issue fetching videos
 
     logger.info(`Found ${videos.length} videos`);
 
     await deleteVideosByChannelId(channel.channel_id); // Clean up old videos in case any were made private
-    for (const video of videos) await upsertVideo(video); // Insert all identified videos
-    logger.info("-------------------------------------------------------------------");
+    for (const video of videoApiResponseToDbColumns(videos))
+      await upsertVideo(video); // Insert all identified videos
+    logger.info(
+      "-------------------------------------------------------------------"
+    );
   }
 
   await upsertLastStoreAllVideosDate(new Date());
 };
 
-const getAllPlaylistAsmrVideos = async (
+const getAllPlaylistAsmrVideos = async (channelUploadsPlaylistId) => {
+  try {
+    return await getAllPlaylistAsmrVideosRecusive(channelUploadsPlaylistId);
+  } catch (error) {
+    logger.error(error);
+    return null;
+  }
+};
+
+const getAllPlaylistAsmrVideosRecusive = async (
   channelUploadsPlaylistId,
   nextPageToken = null,
   previousVideos = []
 ) => {
-  const { videos, nextPageToken: newNextPageToken } = await getAsmrVideosInPlaylist(
-    channelUploadsPlaylistId,
-    nextPageToken
-  );
+  const { videos, nextPageToken: newNextPageToken } =
+    await getAsmrVideosInPlaylist(channelUploadsPlaylistId, nextPageToken);
 
   const allVideos = [...previousVideos, ...videos];
   if (!newNextPageToken) return allVideos;
 
   // Search again with next page token if one is available
-  return getAllPlaylistAsmrVideos(channelUploadsPlaylistId, newNextPageToken, allVideos);
+  return getAllPlaylistAsmrVideosRecusive(
+    channelUploadsPlaylistId,
+    newNextPageToken,
+    allVideos
+  );
 };
